+++ cli.js
@@ -2,8 +2,10 @@
 let create = require('.')
 let banner = require('./src/bootstrap/_banner')
 let bootstrap = require('./src/bootstrap')
+let getName = require('./src/bootstrap/_get-name')
 let { version } = require('./package.json')
 let { updater } = require('@architect/utils')
+let _inventory = require('@architect/inventory')
 let simpleStatic = require('./src/simple-static')
 let update = updater('Create')
 
@@ -24,12 +26,16 @@ let isStatic =  opt => opt === 'static' || opt === '--static' || opt === '-s'
 
 // eslint-disable-next-line
 async function cmd () {
-  // Print banner
-  banner({ version })
-
   // Used by bootstrap to differentiate between arc create and preflight bootstrap calls
   let standalone = true
   let options = process.argv
+
+  // Get the folder name so we know where to inventory
+  let { folder } = getName({ options, update })
+  let inventory = await _inventory({ cwd: folder })
+
+  // Print banner
+  banner({ inventory, version })
 
   if (options.some(isStatic)) {
     // just bail early here ... I don't use understand the code below so just stubbing in
@@ -43,7 +49,7 @@ async function cmd () {
   }
 
   // Bootstrap the project on the filesystem, including new dirs, npm i, etc.
-  let { folder, install } = bootstrap({ options, standalone, update, runtime: opts.runtime })
+  let { install } = bootstrap({ options, standalone, update, runtime: opts.runtime })
 
   return create({ options: opts, folder, install, standalone, update })
 }