+++ src/lambda/index.js
@@ -1,10 +1,9 @@
 let { existsSync, mkdirSync } = require('fs')
 let { join } = require('path')
 let parallel = require('run-parallel')
-let { getLambdaName } = require('@architect/utils')
-
 let getExtension = require('./get-extension')
 let writeArcConfig = require('./write-arc-config')
+let writeTemplate = require('./write-template')
 let writeCode = require('./write-code')
 
 /**
@@ -14,40 +13,44 @@ let writeCode = require('./write-code')
  * @param {string} path - optional. of the format: /foo/:bar/baz
  * @param {string} name - optional. of the format: hello-world-2020
  */
-module.exports = function code ({ type, runtime, method, path, name, folder }, callback) {
+module.exports = function code (params, callback) {
+  let { arcStaticAssetProxy, src, type, runtime, prefs } = params
 
-  let dest = type === 'http' ? `${method.toLowerCase()}${getLambdaName(path)}` : name
-  let extension = getExtension(runtime)
-  let basePath = join(folder, 'src', type, dest)
-  let fullPath = join(basePath, `index.${extension}`)
-  let configPath = join(basePath, `.arc-config`)
-
-  // immediate bail if either file exists
-  if (existsSync(fullPath) || existsSync(configPath)) {
+  // Immediate bail if source dir already exists
+  if (arcStaticAssetProxy || existsSync(src)) {
     callback()
   }
   else {
-    mkdirSync(basePath, { recursive: true })
+    let handlerFile = join(src, `index.${getExtension(runtime)}`)
+    mkdirSync(src, { recursive: true })
 
     parallel({
       config (callback) {
         writeArcConfig({
-          configPath,
+          configPath: join(src, `config.arc`),
           runtime
         }, callback)
       },
       code (callback) {
-        writeCode({
-          type,
-          runtime,
-          fullPath
-        }, callback)
+        let templates = prefs && prefs.create && prefs.create.templates
+        let template = templates && templates[type]
+        if (template) {
+          writeTemplate({
+            template,
+            handlerFile
+          }, callback)
+        }
+        else {
+          writeCode({
+            type,
+            runtime,
+            handlerFile
+          }, callback)
+        }
       }
     }, function done (err) {
       if (err) callback(err)
-      else {
-        callback(null, `src/${type}`)
-      }
+      else callback(null, src)
     })
   }
 }