+++ changelog.md
@@ -2,6 +2,27 @@
 
 ---
 
+## [1.3.0] 2020-10-20
+
+### Added
+
+- Added support for custom file paths
+- Added support for custom default templates, thanks @Ankcorn!
+  - Add per-pragma custom templates via the Arc preferences file, example:
+```arc
+# preferences.arc or prefs.arc
+@create
+templates
+  http path/to/template/file
+```
+
+
+### Changed
+
+- Implemented Inventory (`@architect/inventory`)
+
+---
+
 ## [1.2.0] 2020-09-30
 
 ### Added